Still display email auth link when OmniAuth identity exists
Currently, because `allow_email_auth?` is set to false when an OmniAuth
identity is linked for an account, entering valid email on the login
page doesn't display the email auth link. This is not desirable, we
still want to normally allow email auth if OmniAuth is setup.

To fix this, rather than preventing `email_auth` from being added to
`#possible_authentication_methods` when OmniAuth is configured, we
instead prevent adding `omniauth` if either `password` or `email_auth`
would already be added.